+++ Package.swift
@@ -18,7 +18,7 @@ let package = Package(
     dependencies: [
         .package(
             url: "https://github.com/open-telemetry/opentelemetry-swift-core.git",
-            exact: "2.1.1"
+            exact: "2.2.0"
         ),
         .package(
             url: "https://github.com/open-telemetry/opentelemetry-swift.git",