+++ Sources/Honeycomb/Session/SessionStorage.swift
@@ -1,6 +1,7 @@
 import Foundation
+import OpenTelemetryApi
 
-private let sessionIdKey: String = "session.id"
+private let sessionIdKey: String = SemanticConventions.Session.id.rawValue
 private let sessionStartTimeKey: String = "session.startTime"
 
 // A set of utility functions for reading and saving a Session object to persistent storage